# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/common/file/avatar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上传用户头像图片",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common"
                ],
                "summary": "上传头像",
                "parameters": [
                    {
                        "type": "file",
                        "description": "头像图片",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功,返回上传后的文件名",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "上传失败",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
        "/common/file/video": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上传用户视频",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common"
                ],
                "summary": "上传视频",
                "parameters": [
                    {
                        "type": "file",
                        "description": "视频文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "Access-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功,返回上传后的文件名",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/resp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "上传失败",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "resp.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "default": 1000
                },
                "msg": {
                    "type": "string"
                },
                "time_now": {
                    "type": "integer"
                }
            }
        },
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "default": 0
                },
                "data": {},
                "time_now": {
                    "type": "integer"
                }
            }
        },
...
    }
}